requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
env==0.1.0
environ==1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
